[Issue #80] Group Refiner file remux job family
