+++ apps/backend/src/mediamop/modules/refiner/file_remux_pass/job_kinds.py
@@ -1,0 +1,5 @@
+"""Durable Refiner job kinds: per-file remux pass (ffprobe + plan + optional ffmpeg)."""
+
+from __future__ import annotations
+
+REFINER_FILE_REMUX_PASS_JOB_KIND = "refiner.file.remux_pass.v1"
